extract logic for binary generation outside of predicate.rs
